Unified LivechatInquiry in one file
