--- packages/rocketchat-livechat/client/collections/LivechatInquiry.js
@@ -1,3 +1,0 @@
-import { Mongo } from 'meteor/mongo';
-
-export const LivechatInquiry = new Mongo.Collection('rocketchat_livechat_inquiry');
