+++ packages/rocketchat-livechat/client/roomType.js
@@ -1,4 +1,4 @@
 import { RocketChat } from 'meteor/rocketchat:lib';
-import LivechatRoomType from './lib/LivechatRoomType';
+import LivechatRoomType from '../lib/LivechatRoomType';
 
 RocketChat.roomTypes.add(new LivechatRoomType());
